moved the PageNotFound component on pages folder
