+++ src/pages/PageNotFound.jsx
@@ -1,47 +1,23 @@
-// import styled from "styled-components";
+import { useNavigate } from "react-router-dom";
+import styled from "styled-components";
 
-// import { useMoveBack } from "../hooks/useMoveBack";
-// import Heading from "../ui/Heading";
+const PageNotFoundWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  height: 100vh;
+  gap: 2em;
+`;
 
-// const StyledPageNotFound = styled.main`
-//   height: 100vh;
-//   background-color: var(--color-grey-50);
-//   display: flex;
-//   align-items: center;
-//   justify-content: center;
-//   padding: 4.8rem;
-// `;
+function PageNotFound() {
+  const navigate = useNavigate();
+  return (
+    <PageNotFoundWrapper>
+      <p>| Page cannot be displayed</p>
+      <button onClick={() => navigate(-1)}>&larr;go back</button>
+    </PageNotFoundWrapper>
+  );
+}
 
-// const Box = styled.div`
-//   /* box */
-//   background-color: var(--color-grey-0);
-//   border: 1px solid var(--color-grey-100);
-//   border-radius: var(--border-radius-md);
-
-//   padding: 4.8rem;
-//   flex: 0 1 96rem;
-//   text-align: center;
-
-//   & h1 {
-//     margin-bottom: 3.2rem;
-//   }
-// `;
-
-// function PageNotFound() {
-//   const moveBack = useMoveBack();
-
-//   return (
-//     <StyledPageNotFound>
-//       <Box>
-//         <Heading as="h1">
-//           The page you are looking for could not be found 😢
-//         </Heading>
-//         <button onClick={moveBack} size="large">
-//           &larr; Go back
-//         </button>
-//       </Box>
-//     </StyledPageNotFound>
-//   );
-// }
-
-// export default PageNotFound;
+export default PageNotFound;
